feat(ssr-until-sw): cleanup
